fix: drop iron-flex-layout
Dropped the iron-flex-layout dependency and tried redefining missing
styles in the proper contexts. A lot of styles seemed unnecessary from
the demo- and test standpoint, let's see if it remains so in production/FE.
Otherwise we should add tests.

Rewrote the demo with haunted since current version failed for me.

Upgraded dependencies.

Signed-off-by: Patrik Kullman <patrik.kullman@neovici.se>