feat: renderItem: render templates with lit-html
This adds support to pass data-nav a `renderItem` function,
which should return a lit-html TemplateResult when invoked.
If set, this will use lit-html rendering instead of the current
Polymer template handling.

The resize-forwarding behavior of iron-resize has been disabled
when using renderItem. If resizing issues occur when migrating
to renderItem, switch to ResizeObserver in the data-nav view.