feat(cosmoz-data-nav-pure): haunted implementation
I decided to have a whack at a more declarative data-nav. The result is a very reliable data-nav
that does away with all of the complexity of keeping an item cache and mutable data. It relies on
immutable data being passed from the component above. It keeps the maintain selection and need-data
behaviour.

It's missing the transition animation and the incomplete template, which I will add in a follow-up
PR.

This is meant to be used only with renderItem functions that render haunted views. This is not
tested with polymer views and will likely cause bugs, due to polymer's use of mutable state.